requirement.txt
Flask
gspread
google-auth
pandas
matplotlib
pytest
dash
plotly
